Use new image (LXC) Check distribution for network config

+++ ext/sample-config/qvd-sample-init.sh
@@ -8,41 +8,47 @@ host=`hostname`
 
 set -x
 
-if [ ! -f /var/lib/qvd/storage/staging/demo-ubuntu-kde.img ] ; then
-	mkdir -p /var/lib/qvd/storage/staging/
-	file=/var/lib/qvd/storage/staging/demo-ubuntu-kde.img.gz
-	wget http://theqvd.com/downloads/demo/demo-ubuntu-kde.img.gz -c -O "$file" || rm -f "$file"
-	gunzip "$file" || rm -f /var/lib/qvd/storage/staging/*
 
-	if [ ! -f /var/lib/qvd/storage/staging/demo-ubuntu-kde.img ] ; then
+staging=/var/lib/qvd/storage/staging/
+image=ubuntu-12.04-desktop.i386.tar.gz
+url=http://theqvd.com/downloads/qvd3/images/lxc/$image
+
+if [ ! -f "$staging/$image" ] ; then
+	mkdir -p "$staging"
+	compressed="$staging/$image"
+
+	wget "$url" -c -O "$compressed" || rm -f "$compressed"
+
+	if [ ! -f "$staging/$image" ] ; then
 		echo "Failed to download demo image, aborting"
 		exit 1
 	fi
 fi
 
-if [ ! -f /etc/sysconfig/network/ifcfg-qvdnet ] ; then
-	if [ ! -f /etc/sysconfig/network/ifcfg-eth0 ] ; then
-		echo "Can't open network configuration for eth0, aborting"
-		exit 1
-	fi
+if [ -f "/etc/SuSE-release" ] ; then
+	if [ ! -f /etc/sysconfig/network/ifcfg-qvdnet ] ; then
+		if [ ! -f /etc/sysconfig/network/ifcfg-eth0 ] ; then
+			echo "Can't open network configuration for eth0, aborting"
+			exit 1
+		fi
 
-	. /etc/sysconfig/network/ifcfg-eth0
+		. /etc/sysconfig/network/ifcfg-eth0
 
-	# Backup original config file only once -- this backs up the system default file
-	if [ ! -f /etc/sysconfig/network/ifcfg-eth0.orig ] ; then
-		cp -f /etc/sysconfig/network/ifcfg-eth0 /etc/sysconfig/network/ifcfg-eth0.orig
-	fi
+		# Backup original config file only once -- this backs up the system default file
+		if [ ! -f /etc/sysconfig/network/ifcfg-eth0.orig ] ; then
+			cp -f /etc/sysconfig/network/ifcfg-eth0 /etc/sysconfig/network/ifcfg-eth0.orig
+		fi
 
-	# Backup previous config file.
-	cp -f /etc/sysconfig/network/ifcfg-eth0 /etc/sysconfig/network/ifcfg-eth0.bak
+		# Backup previous config file.
+		cp -f /etc/sysconfig/network/ifcfg-eth0 /etc/sysconfig/network/ifcfg-eth0.bak
 
-	if [ -f /etc/sysconfig/network/ifcfg-qvdnet ] ; then
-		cp -f /etc/sysconfig/network/ifcfg-qvdnet /etc/sysconfig/network/ifcfg-qvdnet.bak
-	fi
+		if [ -f /etc/sysconfig/network/ifcfg-qvdnet ] ; then
+			cp -f /etc/sysconfig/network/ifcfg-qvdnet /etc/sysconfig/network/ifcfg-qvdnet.bak
+		fi
 
-	# Generate new network config
-	
-	cat >/etc/sysconfig/network/ifcfg-qvdnet <<CONF
+		# Generate new network config
+		
+		cat >/etc/sysconfig/network/ifcfg-qvdnet <<CONF
 BOOTPROTO='$BOOTPROTO'
 GATEWAY='$GATEWAY'
 IPADDR='$IPADDR'
@@ -55,7 +61,7 @@ BRIDGE_STP='on'
 NAME='QVD Network'
 CONF
 
-	cat >/etc/sysconfig/network/ifcfg-eth0 <<CONF2
+		cat >/etc/sysconfig/network/ifcfg-eth0 <<CONF2
 BOOTPROTO='static'
 BROADCAST=''
 IPADDR='0.0.0.0'
@@ -65,9 +71,13 @@ NETWORK=''
 USERCONTROL='no'
 CONF2
 
-	service network restart
+		service network restart
+	fi
+elif [ -f /etc/debian_version ] ; then
+	echo "Debian support not yet complete"
+else
+	echo "Can't configure network, unknown distribution"
 fi
-
 
 qvd-deploy-db.pl --force 2>&1 | tee /tmp/sample-init.log || exit 1;
 
@@ -77,7 +87,7 @@ qvd-admin.pl config set vm.network.ip.start=10.1.0.30
 qvd-admin.pl host add name=$host address=127.0.0.1
 
 qvd-admin.pl osf  add name=sample
-qvd-admin.pl di   add osf_id=1 path=/var/lib/qvd/storage/staging/demo-ubuntu-kde.img
+qvd-admin.pl di   add osf_id=1 path=$staging/$image
 qvd-admin.pl user add login=qvd password=qvd
 qvd-admin.pl vm   add name=sample_vm user=qvd osf=sample
